pyproject.toml
[project]
name = "konatagger"
version = "0.1.0"
description = "Tagging images API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.7",
    "einops>=0.7.0",
    "fastapi>=0.115.7",
    "pillow>=9.4.0",
    "python-multipart>=0.0.20",
    "safetensors>=0.4.1",
    "torch>=2.0.1,<3.0.0",
    "torchvision>=0.15.2",
    "transformers>=4.36.2",
    "uvicorn>=0.34.0",
]
